requirements.txt
# Frontend
streamlit

# Config and environment
pyyaml
python-dotenv

# Data and modeling
pandas
numpy
cvxpy
scipy
yfinance
requests

# Visualization
matplotlib
plotly
